Add files via upload
Diabetes is a chronic health condition that affects millions of people worldwide. Early detection and prediction of diabetes can significantly contribute to effective management and prevention of complications. In this context, the goal is to develop a binary classification model that predicts whether a person is likely to have diabetes based on certain features.